Make ACL creation failure explicit when user is an owner
Why:
We are working to improve the user experience around managing ACLs. With this
change we will make it clearer that it is failing to be created because the user
is an owner (and therefore ACL control is managed by the CustomAclAuthorizer, not
Zookeeper)

This introduces a property `kas.authorizer.owners` which will accept a comma
separated list of owner names. The plugin is responsible for appending the principal
type prefix.