Make custom owner ACL rules visible by returning them from acl
Combine owner ACLs with the list returned from Zookeeper so users can see that these
rules exist from kafka tooling.

Note:
To enable kafka to describe CustomAclBinding when building a response we have changed
ApiAwareAccessControlEntry to return an operation. If it has a single, non-excluded
operation in it's operations we return that else return UNKNOWN.