Delete orphaned ACL bindings for managed principals
Why:
To avoid misleading users we delete principals that are explicitly targeted by CustomAclAuthorizer rules.
It is possible for a user with ACL rules in Kafka's Control Plane to become a user managed by CustomAclAuthorizer,
in which case we would never use or update rules for them in Kafka's Control Plane. The old rules remain, dangling,
in Kafka's Control Plane and visible through kafka ACL tooling. This would mislead users because there will be a
mismatch between what the Kafka Control Plane rules say they should be permitted to do and the truth.

MGDSTRM-9583